Make SQLAlchemy-based test data setup available to all tests
